pyproject.toml
[project]
name = "comfyui-light-tool"
description = "An awesome light image processing tool nodes for ComfyUI."
version = "1.0.7"
license = { file = "LICENSE" }
dependencies = ["numpy", "pillow", "torch", "opencv-python", "httpx", "oss2"]

[project.urls]
Repository = "https://github.com/ihmily/ComfyUI-Light-Tool"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "hmily"
DisplayName = "Comfyui-Light-Tool"
Icon = ""